+++ app/(normalLayout)/mon-compte/(subpages)/layout.tsx
@@ -6,11 +6,16 @@ import { SideMenu } from '@codegouvfr/react-dsfr/SideMenu';
 // Auth
 import { signOut } from 'next-auth/react';
 
+// Nav & routes
+import { usePathname } from 'next/navigation';
+
 export default function MyAccountLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
+  const pathname = usePathname();
+
   // @ts-ignore
   const handleSignout = (e) => {
     e.preventDefault();
@@ -21,13 +26,14 @@ export default function MyAccountLayout({
     <>
       <div className="fr-container">
         <div className="fr-grid-row fr-grid-row--gutters fr-py-12v">
-          <div className="fr-col-4">
+          <div className="fr-col-3">
             <SideMenu
               align="left"
               burgerMenuButtonText="Dans cette rubrique"
               items={[
                 {
-                  text: 'Mes clés API',
+                  text: "Mes clés d'API",
+                  isActive: pathname === '/mon-compte/cles-api',
                   linkProps: {
                     href: '/mon-compte/cles-api',
                   },
@@ -46,7 +52,7 @@ export default function MyAccountLayout({
               ]}
             />
           </div>
-          <div className="fr-col-8">{children}</div>
+          <div className="fr-col-9">{children}</div>
         </div>
       </div>
     </>
