+++ app/(normalLayout)/mon-compte/page.tsx
@@ -1,13 +1,7 @@
-export default function MyAccount() {
-  return (
-    <>
-      <div className="fr-container">
-        <div className="fr-grid-row">
-          <div className="fr-col-12">
-            <h1>Mon compte</h1>
-          </div>
-        </div>
-      </div>
-    </>
-  );
+'use server';
+
+import { redirect } from 'next/navigation';
+
+export default async function Page() {
+  redirect('/mon-compte/cles-api');
 }
